Banished playback fighting with Kodi.
Latency still seems to be biting us, but it's marginally usable. Any
client but Kodi pausing immediately gets reset, and I see in debug
output that a "playing" state is reported by the player after syncplay
logs reception of the pause command. So there's still some problem with
RPC latency.